+++ src/web/hooks/useStripeCustomerPortal.js
@@ -1,0 +1,3 @@
+export const useStripeCustomerPortal = () => {
+    
+}
